Create getObjectList.R
New function for getting object lists, didn't change too much from @ljvillanueva's getContent function and it works for me but would love a second eye in case  something is amiss.

I also have never added something to a package before so I assume roxygen would need to be run again to add the function to the NAMESPACE but don't quite know how to do that.  Thanks!